+++ Classes/Mvc/Cli/Command.php
@@ -14,11 +14,12 @@ namespace Helhum\Typo3Console\Mvc\Cli;
  *
  */
 
+use Doctrine\Common\Annotations\AnnotationReader;
 use Helhum\Typo3Console\Mvc\Cli\Symfony\Application;
 use Symfony\Component\Console\Exception\InvalidArgumentException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
-use TYPO3\CMS\Extbase\Reflection\MethodReflection;
+use TYPO3\CMS\Extbase\Reflection\ClassSchema;
 
 /**
  * Represents a Command
@@ -39,11 +40,6 @@ class Command
      * @var string
      */
     protected $commandIdentifier;
-
-    /**
-     * @var MethodReflection
-     */
-    protected $commandMethodReflection;
 
     /**
      * Name of the extension to which this command belongs
@@ -88,6 +84,16 @@ class Command
     private $inputDefinitions;
 
     /**
+     * @var ClassSchema
+     */
+    private $classSchema;
+
+    /**
+     * @var string
+     */
+    private $controllerCommandMethod;
+
+    /**
      * @param \TYPO3\CMS\Extbase\Reflection\ReflectionService $reflectionService
      */
     public function injectReflectionService(\TYPO3\CMS\Extbase\Reflection\ReflectionService $reflectionService)
@@ -102,8 +108,12 @@ class Command
      */
     public function __construct(string $controllerClassName, string $controllerCommandName)
     {
+        // For now (before switching to Doctrine Annotations) we ignore this annotation
+        AnnotationReader::addGlobalIgnoredName('definition');
+
         $this->controllerClassName = $controllerClassName;
         $this->controllerCommandName = $controllerCommandName;
+        $this->controllerCommandMethod = $this->controllerCommandName . 'Command';
         $delimiter = strpos($controllerClassName, '\\') !== false ? '\\' : '_';
         $classNameParts = explode($delimiter, $controllerClassName);
         if (isset($classNameParts[0]) && $classNameParts[0] === 'TYPO3' && isset($classNameParts[1]) && $classNameParts[1] === 'CMS') {
@@ -127,6 +137,11 @@ class Command
         $this->extensionName = $classNameParts[1];
         $extensionKey = \TYPO3\CMS\Core\Utility\GeneralUtility::camelCaseToLowerCaseUnderscored($this->extensionName);
         $this->commandIdentifier = strtolower($extensionKey . ':' . substr($classNameParts[$numberOfClassNameParts - 1], 0, -17) . ':' . $controllerCommandName);
+    }
+
+    public function initializeObject()
+    {
+        $this->classSchema = $this->reflectionService->getClassSchema($this->controllerClassName);
     }
 
     /**
@@ -172,7 +187,7 @@ class Command
      */
     public function getShortDescription(): string
     {
-        $lines = explode(LF, $this->getCommandMethodReflection()->getDescription());
+        $lines = explode(LF, $this->classSchema->getMethod($this->controllerCommandMethod)['description']);
         return !empty($lines) ? trim($lines[0]) : '<no description available>';
     }
 
@@ -185,7 +200,7 @@ class Command
      */
     public function getDescription(): string
     {
-        $lines = explode(LF, $this->getCommandMethodReflection()->getDescription());
+        $lines = explode(LF, $this->classSchema->getMethod($this->controllerCommandMethod)['description']);
         array_shift($lines);
         $descriptionLines = [];
         foreach ($lines as $line) {
@@ -204,7 +219,7 @@ class Command
      */
     public function hasArguments(): bool
     {
-        return !empty($this->getCommandMethodReflection()->getParameters());
+        return !empty($this->classSchema->getMethod($this->controllerCommandMethod)['params']);
     }
 
     /**
@@ -270,13 +285,12 @@ class Command
             return $this->argumentDefinitions = [];
         }
         $this->argumentDefinitions = [];
-        $commandMethodReflection = $this->getCommandMethodReflection();
-        $annotations = $commandMethodReflection->getTagsValues();
-        $commandParameters = $this->reflectionService->getMethodParameters($this->controllerClassName, $this->controllerCommandName . 'Command');
+        $commandParameters = $this->classSchema->getMethod($this->controllerCommandMethod)['params'];
+        $commandParameterTags = $this->classSchema->getMethod($this->controllerCommandMethod)['tags']['param'];
         $i = 0;
         $argumentNames = $this->getDefinedArgumentNames();
         foreach ($commandParameters as $commandParameterName => $commandParameterDefinition) {
-            $explodedAnnotation = preg_split('/\s+/', $annotations['param'][$i], 3);
+            $explodedAnnotation = preg_split('/\s+/', $commandParameterTags[$i], 3);
             $description = !empty($explodedAnnotation[2]) ? $explodedAnnotation[2] : '';
             $dataType = $commandParameterDefinition['type'] ?? 'null';
             if ($commandParameterDefinition['array']) {
@@ -398,8 +412,7 @@ class Command
             return $this->commandMethodDefinitions;
         }
         $this->commandMethodDefinitions = [];
-        $commandMethodReflection = $this->getCommandMethodReflection();
-        $annotations = $commandMethodReflection->getTagsValues();
+        $annotations = $this->classSchema->getMethod($this->controllerCommandMethod)['tags'];
         if (!empty($annotations['definition'])) {
             $this->commandMethodDefinitions = $this->parseDefinitions($annotations['definition']);
         }
@@ -407,7 +420,7 @@ class Command
     }
 
     /**
-     * Very simple parsing of definition annotations on command methods.
+     * Very simple parsing of a definition annotations on command methods.
      *
      * @param array $definitionAnnotations Definition tags on command controller command methods
      * @throws InvalidArgumentException
@@ -483,7 +496,7 @@ class Command
      */
     public function isInternal(): bool
     {
-        return $this->getCommandMethodReflection()->isTaggedWith('internal');
+        return isset($this->classSchema->getMethod($this->controllerCommandMethod)['tags']['internal']);
     }
 
     /**
@@ -493,7 +506,7 @@ class Command
      */
     public function isCliOnly(): bool
     {
-        return $this->getCommandMethodReflection()->isTaggedWith('cli');
+        return isset($this->classSchema->getMethod($this->controllerCommandMethod)['tags']['cli']);
     }
 
     /**
@@ -505,7 +518,7 @@ class Command
      */
     public function isFlushingCaches(): bool
     {
-        return $this->getCommandMethodReflection()->isTaggedWith('flushesCaches');
+        return isset($this->classSchema->getMethod($this->controllerCommandMethod)['tags']['flushesCaches']);
     }
 
     /**
@@ -517,28 +530,15 @@ class Command
      */
     public function getRelatedCommandIdentifiers(): array
     {
-        $commandMethodReflection = $this->getCommandMethodReflection();
-        if (!$commandMethodReflection->isTaggedWith('see')) {
+        if (!isset($this->classSchema->getMethod($this->controllerCommandMethod)['tags']['see'])) {
             return [];
         }
         $relatedCommandIdentifiers = [];
-        foreach ($commandMethodReflection->getTagValues('see') as $tagValue) {
+        foreach ($this->classSchema->getMethod($this->controllerCommandMethod)['tags']['see'] as $tagValue) {
             if (preg_match('/^[\\w\\d\\.]+:[\\w\\d]+:[\\w\\d]+$/', $tagValue) === 1) {
                 $relatedCommandIdentifiers[] = $tagValue;
             }
         }
         return $relatedCommandIdentifiers;
-    }
-
-    /**
-     * @throws \ReflectionException
-     * @return MethodReflection
-     */
-    protected function getCommandMethodReflection(): MethodReflection
-    {
-        if ($this->commandMethodReflection === null) {
-            $this->commandMethodReflection = new MethodReflection($this->controllerClassName, $this->controllerCommandName . 'Command');
-        }
-        return $this->commandMethodReflection;
     }
 }
