+++ composer.json
@@ -56,6 +56,9 @@
             "Helhum\\Typo3Console\\": [
                 "Classes/",
                 "Resources/Private/ExtensionArtifacts/src/"
+            ],
+            "Helhum\\Typo3Console\\TYPO3v87\\": [
+                "Compatibility/TYPO3v87/"
             ]
         }
     },
